requirements.txt
opencv-python
requests
python-dotenv
